% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_scientific_names.R
\name{match_NTC_names}
\alias{match_NTC_names}
\title{Match Scientific Names}
\usage{
match_NTC_names(names)
}
\arguments{
\item{names}{A character vector of scientific names to be matched.}
}
\value{
A data frame with the original names and a logical indicating if they were matched.
}
\description{
This function matches user-provided scientific names with a predefined list of scientific names.
}
